Added foundation for testing
added Maven and Junit, declared the test src and resources and made a carfactory test
