+++ Design Pattern/src/main/java/LuxuryCar.java
