+++ Design Pattern/src/main/java/Main.java
