+++ Design Pattern/src/main/java/RentalAgreement.java
@@ -18,14 +18,11 @@ class RentalAgreement implements RentalAgreementInterface {
     private CarInterface car;
     private int duration;
     private String customer;
-    private int initialOdometerReading;
-    private int finalOdometerReading;
 
-    public RentalAgreement(CarInterface car, int duration, String customer, int initialOdometerReading) {
+    public RentalAgreement(CarInterface car, int duration, String customer) {
         this.car = car;
         this.duration = duration;
         this.customer = customer;
-        this.initialOdometerReading = initialOdometerReading;
     }
 
     public String make_rental_agreement() {
@@ -35,11 +32,4 @@ class RentalAgreement implements RentalAgreementInterface {
     public int getDuration() {
         return duration;
     }
-
-    public int returnCar(int finalOdometerReading) {
-        this.finalOdometerReading = finalOdometerReading;
-        int distanceDriven = finalOdometerReading - initialOdometerReading;
-        return distanceDriven;
-    }
 }
-
