+++ Design Pattern/src/main/java/StandardCar.java
