+++ Design Pattern/src/main/java/TowbarDecorator.java
