+++ Design Pattern/src/main/java/TripleADecorator.java
