--- DesignPatterns/src/RentalCompany.java
@@ -1,6 +1,0 @@
-import java.util.List;
-public class RentalCompany {
-    private String location;
-    private String name;
-    private List<Car>;
-}
